Split llama_cpp/lib.rs into smaller files
